+++ base/ChangeLog.md
@@ -11,6 +11,8 @@
 
 - Add support for PPC32 and PPC64 relocations in `Data.Macaw.Memory.ElfLoader`.
 
+- Add support for RISC-V relocations in `Data.Macaw.Memory.ElfLoader`.
+
 ### API Changes
 
 - Architecture-specific block terminators can now contain macaw values